fix: correcoes de qualidade e bugs no backend
- Locking otimista (@Version) em Produto para evitar race condition no estoque
- Endpoint DELETE /vendas/{id} exposto no VendaController
- Valores monetarios migrados de double para BigDecimal (preco, precoUnit, total)
- ProdutoRepository: parametros de faixa de preco migrados para BigDecimal, metodo List sem paginacao removido
- ProdutoService/ProdutoController: faixa de preco migrada para BigDecimal
- GlobalExceptionHandler: handler para ObjectOptimisticLockingFailureException
- SecurityConfig: @Lazy nos filtros JWT e RateLimit para resolver referencia circular
- VendaServiceTest: corrigidos testes com assinatura errada apos migracao BigDecimal

+++ src/main/java/com/loja/movapp/controller/ProdutoController.java
@@ -15,6 +15,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.net.URI;
 
 @RestController
@@ -93,8 +94,8 @@ public class ProdutoController {
     @Operation(summary = "Buscar por faixa de preço",
             description = "Retorna produtos entre o preço mínimo e máximo com paginação")
     public ResponseEntity<Page<ProdutoResponseDTO>> buscarPorFaixaDePreco(
-            @Parameter(description = "Preço mínimo") @RequestParam double min,
-            @Parameter(description = "Preço máximo") @RequestParam double max,
+            @Parameter(description = "Preço mínimo") @RequestParam BigDecimal min,
+            @Parameter(description = "Preço máximo") @RequestParam BigDecimal max,
             Pageable pageable) {
         Page<ProdutoResponseDTO> produtos = service.buscarPorFaixaDePreco(min, max, pageable);
         if (produtos.isEmpty()) {
