+++ src/main/java/com/loja/movapp/dto/ProdutoResponseDTO.java
@@ -1,16 +1,18 @@
 package com.loja.movapp.dto;
 
+import java.math.BigDecimal;
+
 public class ProdutoResponseDTO {
 
     private String codigo;
     private String nome;
     private String cor;
     private String tamanho;
-    private double preco;
+    private BigDecimal preco;
     private int estoque;
 
     public ProdutoResponseDTO(String codigo, String nome, String cor, String tamanho,
-                              double preco, int estoque) {
+                              BigDecimal preco, int estoque) {
 
         this.codigo  = codigo;
         this.nome    = nome;
@@ -20,10 +22,10 @@ public class ProdutoResponseDTO {
         this.estoque = estoque;
     }
 
-    public String getCodigo()  {return codigo; }
-    public String getNome()    {return nome; }
-    public String getCor()     {return cor; }
-    public String getTamanho() {return tamanho; }
-    public double getPreco()   {return preco; }
-    public int getEstoque()    {return estoque; }
+    public String    getCodigo()  {return codigo; }
+    public String    getNome()    {return nome; }
+    public String    getCor()     {return cor; }
+    public String    getTamanho() {return tamanho; }
+    public BigDecimal getPreco()  {return preco; }
+    public int       getEstoque() {return estoque; }
 }
